refactor: BuyXGetY 할인 금액 도출 함수 분리
